requirements.txt
psutil>=5.8.0
aiohttp>=3.8.0
python-telegram-bot>=13.7
pytest>=7.0.0
pytest-asyncio>=0.18.0
python-dotenv>=0.19.0
coverage>=6.3.0
pytest-cov>=2.12.1
numpy>=1.21.0
shapely>=2.0.0
folium>=0.14.0
cryptography>=3.4.0
matplotlib>=3.5.0
seaborn>=0.12.0
gpxpy>=1.5.0
geopy>=2.3.0
polyline>=2.0.0
simplekml>=1.3.1
requests>=2.26.0
aiosqlite>=0.17.0
